gtk-2: set colors for tabwin.
tabwin is the window switcher of XFCE4. The default color scheme makes
the entry pointed by the mouse cursor most prominent, but I'm often
even unaware of where the mouse cursor is when I use the tabwin with
Alt+Tab. So I think the most prominent item should be the one selected
by keyboard.